Fixed imports by moving utils under simulationdataschema/
Fix creating of Symmetry and ChemicalFormula
